requirements.txt
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
greenlet==2.0.2
idna==3.4
niconico.py==1.2.3
numpy==1.24.2
pandas==1.5.3
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.3
typing-extensions==4.4.0
urllib3==1.26.14
